Remove IPC from SubOnResourceLoaded
The original implementation of the resource loaded script provide a
copy of the page source on the response. As the design of the associated
processes evolved the requirement for this went away.

By removing the requirement to gather the page source, the handler is
significantly simplified by removing the need to post the source from
the renderer process

NOTE: If we ever wanted to restore the IPC behaviour, it should be
      reimplemented using the new IPC provided by CefBaseApp rather
      than raw Message passing

The only change to the tests is to the source from response parsing...